prepare for deployment & prettier

+++ contracts/manifold/ERC721Creator.sol
@@ -13,144 +13,233 @@ import "./core/ERC721CreatorCore.sol";
  * @dev ERC721Creator implementation
  */
 contract ERC721Creator is AdminControl, ERC721, ERC721CreatorCore {
-
-    constructor (string memory _name, string memory _symbol) ERC721(_name, _symbol) {
-    }
+    constructor(string memory _name, string memory _symbol)
+        ERC721(_name, _symbol)
+    {}
 
     /**
      * @dev See {IERC165-supportsInterface}.
      */
-    function supportsInterface(bytes4 interfaceId) public view virtual override(ERC721, ERC721CreatorCore, AdminControl) returns (bool) {
-        return ERC721CreatorCore.supportsInterface(interfaceId) || ERC721.supportsInterface(interfaceId) || AdminControl.supportsInterface(interfaceId);
+    function supportsInterface(bytes4 interfaceId)
+        public
+        view
+        virtual
+        override(ERC721, ERC721CreatorCore, AdminControl)
+        returns (bool)
+    {
+        return
+            ERC721CreatorCore.supportsInterface(interfaceId) ||
+            ERC721.supportsInterface(interfaceId) ||
+            AdminControl.supportsInterface(interfaceId);
     }
 
-    function _beforeTokenTransfer(address from, address to, uint256 tokenId) internal virtual override {
-        _approveTransfer(from, to, tokenId);    
+    function _beforeTokenTransfer(
+        address from,
+        address to,
+        uint256 tokenId
+    ) internal virtual override {
+        _approveTransfer(from, to, tokenId);
     }
 
     /**
      * @dev See {ICreatorCore-registerExtension}.
      */
-    function registerExtension(address extension, string calldata baseURI) external override adminRequired nonBlacklistRequired(extension) {
+    function registerExtension(address extension, string calldata baseURI)
+        external
+        override
+        adminRequired
+        nonBlacklistRequired(extension)
+    {
         _registerExtension(extension, baseURI, false);
     }
 
     /**
      * @dev See {ICreatorCore-registerExtension}.
      */
-    function registerExtension(address extension, string calldata baseURI, bool baseURIIdentical) external override adminRequired nonBlacklistRequired(extension) {
+    function registerExtension(
+        address extension,
+        string calldata baseURI,
+        bool baseURIIdentical
+    ) external override adminRequired nonBlacklistRequired(extension) {
         _registerExtension(extension, baseURI, baseURIIdentical);
     }
-
 
     /**
      * @dev See {ICreatorCore-unregisterExtension}.
      */
-    function unregisterExtension(address extension) external override adminRequired {
+    function unregisterExtension(address extension)
+        external
+        override
+        adminRequired
+    {
         _unregisterExtension(extension);
     }
 
     /**
      * @dev See {ICreatorCore-blacklistExtension}.
      */
-    function blacklistExtension(address extension) external override adminRequired {
+    function blacklistExtension(address extension)
+        external
+        override
+        adminRequired
+    {
         _blacklistExtension(extension);
     }
 
     /**
      * @dev See {ICreatorCore-setBaseTokenURIExtension}.
      */
-    function setBaseTokenURIExtension(string calldata uri) external override extensionRequired {
+    function setBaseTokenURIExtension(string calldata uri)
+        external
+        override
+        extensionRequired
+    {
         _setBaseTokenURIExtension(uri, false);
     }
 
     /**
      * @dev See {ICreatorCore-setBaseTokenURIExtension}.
      */
-    function setBaseTokenURIExtension(string calldata uri, bool identical) external override extensionRequired {
+    function setBaseTokenURIExtension(string calldata uri, bool identical)
+        external
+        override
+        extensionRequired
+    {
         _setBaseTokenURIExtension(uri, identical);
     }
 
     /**
      * @dev See {ICreatorCore-setTokenURIPrefixExtension}.
      */
-    function setTokenURIPrefixExtension(string calldata prefix) external override extensionRequired {
+    function setTokenURIPrefixExtension(string calldata prefix)
+        external
+        override
+        extensionRequired
+    {
         _setTokenURIPrefixExtension(prefix);
     }
 
     /**
      * @dev See {ICreatorCore-setTokenURIExtension}.
      */
-    function setTokenURIExtension(uint256 tokenId, string calldata uri) external override extensionRequired {
+    function setTokenURIExtension(uint256 tokenId, string calldata uri)
+        external
+        override
+        extensionRequired
+    {
         _setTokenURIExtension(tokenId, uri);
     }
 
     /**
      * @dev See {ICreatorCore-setTokenURIExtension}.
      */
-    function setTokenURIExtension(uint256[] memory tokenIds, string[] calldata uris) external override extensionRequired {
+    function setTokenURIExtension(
+        uint256[] memory tokenIds,
+        string[] calldata uris
+    ) external override extensionRequired {
         require(tokenIds.length == uris.length, "Invalid input");
-        for (uint i = 0; i < tokenIds.length; i++) {
-            _setTokenURIExtension(tokenIds[i], uris[i]);            
+        for (uint256 i = 0; i < tokenIds.length; i++) {
+            _setTokenURIExtension(tokenIds[i], uris[i]);
         }
     }
 
     /**
      * @dev See {ICreatorCore-setBaseTokenURI}.
      */
-    function setBaseTokenURI(string calldata uri) external override adminRequired {
+    function setBaseTokenURI(string calldata uri)
+        external
+        override
+        adminRequired
+    {
         _setBaseTokenURI(uri);
     }
 
     /**
      * @dev See {ICreatorCore-setTokenURIPrefix}.
      */
-    function setTokenURIPrefix(string calldata prefix) external override adminRequired {
+    function setTokenURIPrefix(string calldata prefix)
+        external
+        override
+        adminRequired
+    {
         _setTokenURIPrefix(prefix);
     }
 
     /**
      * @dev See {ICreatorCore-setTokenURI}.
      */
-    function setTokenURI(uint256 tokenId, string calldata uri) external override adminRequired {
+    function setTokenURI(uint256 tokenId, string calldata uri)
+        external
+        override
+        adminRequired
+    {
         _setTokenURI(tokenId, uri);
     }
 
     /**
      * @dev See {ICreatorCore-setTokenURI}.
      */
-    function setTokenURI(uint256[] memory tokenIds, string[] calldata uris) external override adminRequired {
+    function setTokenURI(uint256[] memory tokenIds, string[] calldata uris)
+        external
+        override
+        adminRequired
+    {
         require(tokenIds.length == uris.length, "Invalid input");
-        for (uint i = 0; i < tokenIds.length; i++) {
-            _setTokenURI(tokenIds[i], uris[i]);            
+        for (uint256 i = 0; i < tokenIds.length; i++) {
+            _setTokenURI(tokenIds[i], uris[i]);
         }
     }
 
     /**
      * @dev See {ICreatorCore-setMintPermissions}.
      */
-    function setMintPermissions(address extension, address permissions) external override adminRequired {
+    function setMintPermissions(address extension, address permissions)
+        external
+        override
+        adminRequired
+    {
         _setMintPermissions(extension, permissions);
     }
 
     /**
      * @dev See {IERC721CreatorCore-mintBase}.
      */
-    function mintBase(address to) public virtual override nonReentrant adminRequired returns(uint256) {
+    function mintBase(address to)
+        public
+        virtual
+        override
+        nonReentrant
+        adminRequired
+        returns (uint256)
+    {
         return _mintBase(to, "");
     }
 
     /**
      * @dev See {IERC721CreatorCore-mintBase}.
      */
-    function mintBase(address to, string calldata uri) public virtual override nonReentrant adminRequired returns(uint256) {
+    function mintBase(address to, string calldata uri)
+        public
+        virtual
+        override
+        nonReentrant
+        adminRequired
+        returns (uint256)
+    {
         return _mintBase(to, uri);
     }
 
     /**
      * @dev See {IERC721CreatorCore-mintBaseBatch}.
      */
-    function mintBaseBatch(address to, uint16 count) public virtual override nonReentrant adminRequired returns(uint256[] memory tokenIds) {
+    function mintBaseBatch(address to, uint16 count)
+        public
+        virtual
+        override
+        nonReentrant
+        adminRequired
+        returns (uint256[] memory tokenIds)
+    {
         tokenIds = new uint256[](count);
         for (uint16 i = 0; i < count; i++) {
             tokenIds[i] = _mintBase(to, "");
@@ -161,9 +250,16 @@ contract ERC721Creator is AdminControl, ERC721, ERC721CreatorCore {
     /**
      * @dev See {IERC721CreatorCore-mintBaseBatch}.
      */
-    function mintBaseBatch(address to, string[] calldata uris) public virtual override nonReentrant adminRequired returns(uint256[] memory tokenIds) {
+    function mintBaseBatch(address to, string[] calldata uris)
+        public
+        virtual
+        override
+        nonReentrant
+        adminRequired
+        returns (uint256[] memory tokenIds)
+    {
         tokenIds = new uint256[](uris.length);
-        for (uint i = 0; i < uris.length; i++) {
+        for (uint256 i = 0; i < uris.length; i++) {
             tokenIds[i] = _mintBase(to, uris[i]);
         }
         return tokenIds;
@@ -172,7 +268,11 @@ contract ERC721Creator is AdminControl, ERC721, ERC721CreatorCore {
     /**
      * @dev Mint token with no extension
      */
-    function _mintBase(address to, string memory uri) internal virtual returns(uint256 tokenId) {
+    function _mintBase(address to, string memory uri)
+        internal
+        virtual
+        returns (uint256 tokenId)
+    {
         _tokenCount++;
         tokenId = _tokenCount;
 
@@ -190,25 +290,45 @@ contract ERC721Creator is AdminControl, ERC721, ERC721CreatorCore {
         return tokenId;
     }
 
-
     /**
      * @dev See {IERC721CreatorCore-mintExtension}.
      */
-    function mintExtension(address to) public virtual override nonReentrant extensionRequired returns(uint256) {
+    function mintExtension(address to)
+        public
+        virtual
+        override
+        nonReentrant
+        extensionRequired
+        returns (uint256)
+    {
         return _mintExtension(to, "");
     }
 
     /**
      * @dev See {IERC721CreatorCore-mintExtension}.
      */
-    function mintExtension(address to, string calldata uri) public virtual override nonReentrant extensionRequired returns(uint256) {
+    function mintExtension(address to, string calldata uri)
+        public
+        virtual
+        override
+        nonReentrant
+        extensionRequired
+        returns (uint256)
+    {
         return _mintExtension(to, uri);
     }
 
     /**
      * @dev See {IERC721CreatorCore-mintExtensionBatch}.
      */
-    function mintExtensionBatch(address to, uint16 count) public virtual override nonReentrant extensionRequired returns(uint256[] memory tokenIds) {
+    function mintExtensionBatch(address to, uint16 count)
+        public
+        virtual
+        override
+        nonReentrant
+        extensionRequired
+        returns (uint256[] memory tokenIds)
+    {
         tokenIds = new uint256[](count);
         for (uint16 i = 0; i < count; i++) {
             tokenIds[i] = _mintExtension(to, "");
@@ -219,17 +339,28 @@ contract ERC721Creator is AdminControl, ERC721, ERC721CreatorCore {
     /**
      * @dev See {IERC721CreatorCore-mintExtensionBatch}.
      */
-    function mintExtensionBatch(address to, string[] calldata uris) public virtual override nonReentrant extensionRequired returns(uint256[] memory tokenIds) {
+    function mintExtensionBatch(address to, string[] calldata uris)
+        public
+        virtual
+        override
+        nonReentrant
+        extensionRequired
+        returns (uint256[] memory tokenIds)
+    {
         tokenIds = new uint256[](uris.length);
-        for (uint i = 0; i < uris.length; i++) {
+        for (uint256 i = 0; i < uris.length; i++) {
             tokenIds[i] = _mintExtension(to, uris[i]);
         }
     }
-    
+
     /**
      * @dev Mint token via extension
      */
-    function _mintExtension(address to, string memory uri) internal virtual returns(uint256 tokenId) {
+    function _mintExtension(address to, string memory uri)
+        internal
+        virtual
+        returns (uint256 tokenId)
+    {
         _tokenCount++;
         tokenId = _tokenCount;
 
@@ -243,7 +374,7 @@ contract ERC721Creator is AdminControl, ERC721, ERC721CreatorCore {
         if (bytes(uri).length > 0) {
             _tokenURIs[tokenId] = uri;
         }
-        
+
         // Call post mint
         _postMintExtension(to, tokenId);
         return tokenId;
@@ -252,7 +383,13 @@ contract ERC721Creator is AdminControl, ERC721, ERC721CreatorCore {
     /**
      * @dev See {IERC721CreatorCore-tokenExtension}.
      */
-    function tokenExtension(uint256 tokenId) public view virtual override returns (address) {
+    function tokenExtension(uint256 tokenId)
+        public
+        view
+        virtual
+        override
+        returns (address)
+    {
         require(_exists(tokenId), "Nonexistent token");
         return _tokenExtension(tokenId);
     }
@@ -261,7 +398,10 @@ contract ERC721Creator is AdminControl, ERC721, ERC721CreatorCore {
      * @dev See {IERC721CreatorCore-burn}.
      */
     function burn(uint256 tokenId) public virtual override nonReentrant {
-        require(_isApprovedOrOwner(msg.sender, tokenId), "Caller is not owner nor approved");
+        require(
+            _isApprovedOrOwner(msg.sender, tokenId),
+            "Caller is not owner nor approved"
+        );
         address owner = ownerOf(tokenId);
         _burn(tokenId);
         _postBurn(owner, tokenId);
@@ -270,14 +410,21 @@ contract ERC721Creator is AdminControl, ERC721, ERC721CreatorCore {
     /**
      * @dev See {ICreatorCore-setRoyalties}.
      */
-    function setRoyalties(address payable[] calldata receivers, uint256[] calldata basisPoints) external override adminRequired {
+    function setRoyalties(
+        address payable[] calldata receivers,
+        uint256[] calldata basisPoints
+    ) external override adminRequired {
         _setRoyaltiesExtension(address(this), receivers, basisPoints);
     }
 
     /**
      * @dev See {ICreatorCore-setRoyalties}.
      */
-    function setRoyalties(uint256 tokenId, address payable[] calldata receivers, uint256[] calldata basisPoints) external override adminRequired {
+    function setRoyalties(
+        uint256 tokenId,
+        address payable[] calldata receivers,
+        uint256[] calldata basisPoints
+    ) external override adminRequired {
         require(_exists(tokenId), "Nonexistent token");
         _setRoyalties(tokenId, receivers, basisPoints);
     }
@@ -285,14 +432,24 @@ contract ERC721Creator is AdminControl, ERC721, ERC721CreatorCore {
     /**
      * @dev See {ICreatorCore-setRoyaltiesExtension}.
      */
-    function setRoyaltiesExtension(address extension, address payable[] calldata receivers, uint256[] calldata basisPoints) external override adminRequired {
+    function setRoyaltiesExtension(
+        address extension,
+        address payable[] calldata receivers,
+        uint256[] calldata basisPoints
+    ) external override adminRequired {
         _setRoyaltiesExtension(extension, receivers, basisPoints);
     }
 
     /**
      * @dev {See ICreatorCore-getRoyalties}.
      */
-    function getRoyalties(uint256 tokenId) external view virtual override returns (address payable[] memory, uint256[] memory) {
+    function getRoyalties(uint256 tokenId)
+        external
+        view
+        virtual
+        override
+        returns (address payable[] memory, uint256[] memory)
+    {
         require(_exists(tokenId), "Nonexistent token");
         return _getRoyalties(tokenId);
     }
@@ -300,7 +457,13 @@ contract ERC721Creator is AdminControl, ERC721, ERC721CreatorCore {
     /**
      * @dev {See ICreatorCore-getFees}.
      */
-    function getFees(uint256 tokenId) external view virtual override returns (address payable[] memory, uint256[] memory) {
+    function getFees(uint256 tokenId)
+        external
+        view
+        virtual
+        override
+        returns (address payable[] memory, uint256[] memory)
+    {
         require(_exists(tokenId), "Nonexistent token");
         return _getRoyalties(tokenId);
     }
@@ -308,7 +471,13 @@ contract ERC721Creator is AdminControl, ERC721, ERC721CreatorCore {
     /**
      * @dev {See ICreatorCore-getFeeRecipients}.
      */
-    function getFeeRecipients(uint256 tokenId) external view virtual override returns (address payable[] memory) {
+    function getFeeRecipients(uint256 tokenId)
+        external
+        view
+        virtual
+        override
+        returns (address payable[] memory)
+    {
         require(_exists(tokenId), "Nonexistent token");
         return _getRoyaltyReceivers(tokenId);
     }
@@ -316,25 +485,42 @@ contract ERC721Creator is AdminControl, ERC721, ERC721CreatorCore {
     /**
      * @dev {See ICreatorCore-getFeeBps}.
      */
-    function getFeeBps(uint256 tokenId) external view virtual override returns (uint[] memory) {
+    function getFeeBps(uint256 tokenId)
+        external
+        view
+        virtual
+        override
+        returns (uint256[] memory)
+    {
         require(_exists(tokenId), "Nonexistent token");
         return _getRoyaltyBPS(tokenId);
     }
-    
+
     /**
      * @dev {See ICreatorCore-royaltyInfo}.
      */
-    function royaltyInfo(uint256 tokenId, uint256 value) external view virtual override returns (address, uint256) {
+    function royaltyInfo(uint256 tokenId, uint256 value)
+        external
+        view
+        virtual
+        override
+        returns (address, uint256)
+    {
         require(_exists(tokenId), "Nonexistent token");
         return _getRoyaltyInfo(tokenId, value);
-    } 
+    }
 
     /**
      * @dev See {IERC721Metadata-tokenURI}.
      */
-    function tokenURI(uint256 tokenId) public view virtual override returns (string memory) {
+    function tokenURI(uint256 tokenId)
+        public
+        view
+        virtual
+        override
+        returns (string memory)
+    {
         require(_exists(tokenId), "Nonexistent token");
         return _tokenURI(tokenId);
     }
-    
 }
